feat: refactor manifest code and add documentation
