+++ pychunkedgraph/app/meshing/common.py
@@ -15,6 +15,7 @@ from pychunkedgraph.app import app_utils
 from pychunkedgraph.graph import chunkedgraph
 from pychunkedgraph.app.meshing import tasks as meshing_tasks
 from pychunkedgraph.meshing import meshgen
+from pychunkedgraph.meshing.manifest import get_highest_child_nodes_with_meshes
 
 
 # -------------------------------
@@ -22,6 +23,7 @@ from pychunkedgraph.meshing import meshgen
 # -------------------------------
 
 __meshing_url_prefix__ = os.environ.get("MESHING_URL_PREFIX", "meshing")
+
 
 def str2bool(v):
     return v.lower() in ("yes", "true", "t", "1")
@@ -153,11 +155,9 @@ def handle_valid_frags(table_id, node_id):
     current_app.user_id = user_id
 
     cg = app_utils.get_cg(table_id)
-
-    seg_ids = meshgen_utils.get_highest_child_nodes_with_meshes(
+    seg_ids = get_highest_child_nodes_with_meshes(
         cg, np.uint64(node_id), stop_layer=1, verify_existence=True
     )
-
     return app_utils.tobinary(seg_ids)
 
 
@@ -188,7 +188,7 @@ def handle_get_manifest(table_id, node_id):
     prepend_seg_ids = request.args.get("prepend_seg_ids", False)
     return_seg_ids = return_seg_ids in ["True", "true", "1", True]
     prepend_seg_ids = prepend_seg_ids in ["True", "true", "1", True]
-    start_layer = cg.get_chunk_layer(np.uint64(node_id))
+    start_layer = cg.meta.custom_data.get("mesh", {}).get("max_layer", 2)
     if "start_layer" in data:
         start_layer = int(data["start_layer"])
 
@@ -209,8 +209,7 @@ def handle_get_manifest(table_id, node_id):
 
 
 def manifest_response(cg, args):
-    from pychunkedgraph.meshing.manifest import speculative_manifest
-    from pychunkedgraph.meshing.manifest import get_highest_child_nodes_with_meshes
+    from pychunkedgraph.meshing.manifest import speculative_manifest_sharded
 
     (
         node_id,
@@ -225,11 +224,10 @@ def manifest_response(cg, args):
     resp = {}
     seg_ids = []
     if not verify:
-        seg_ids, resp["fragments"] = speculative_manifest(cg,
-                                                          node_id,
-                                                          start_layer=start_layer,
-                                                          bounding_box=bounding_box)
-                                                
+        seg_ids, resp["fragments"] = speculative_manifest_sharded(
+            cg, node_id, start_layer=start_layer, bounding_box=bounding_box
+        )
+
     else:
         seg_ids, resp["fragments"] = get_highest_child_nodes_with_meshes(
             cg,
@@ -239,10 +237,8 @@ def manifest_response(cg, args):
             bounding_box=bounding_box,
             flexible_start_layer=flexible_start_layer,
         )
-        if prepend_seg_ids:
-            resp["fragments"] = [
-                f"~{i}:{f}" for i, f in zip(seg_ids, resp["fragments"])
-            ]
+    if prepend_seg_ids:
+        resp["fragments"] = [f"~{i}:{f}" for i, f in zip(seg_ids, resp["fragments"])]
     if return_seg_ids:
         resp["seg_ids"] = seg_ids
     return _check_post_options(cg, resp, data, seg_ids)
@@ -260,23 +256,20 @@ def _check_post_options(cg, resp, data, seg_ids):
     return resp
 
 
-def str2bool(v):
-  return v.lower() in ("yes", "true", "t", "1")
-
 ## REMESHING -----------------------------------------------------
 def handle_remesh(table_id):
     current_app.request_type = "remesh_enque"
     current_app.table_id = table_id
-    is_priority = request.args.get('priority', True, type=str2bool)
-    is_redisjob = request.args.get('use_redis', False, type=str2bool)
+    is_priority = request.args.get("priority", True, type=str2bool)
+    is_redisjob = request.args.get("use_redis", False, type=str2bool)
     user_id = str(g.auth_user["id"])
     current_app.user_id = user_id
 
     new_lvl2_ids = json.loads(request.data)["new_lvl2_ids"]
-    
-    if is_redisjob:    
+
+    if is_redisjob:
         with Connection(redis.from_url(current_app.config["REDIS_URL"])):
-            
+
             if is_priority:
                 retry = Retry(max=3, interval=[1, 10, 60])
                 queue_name = "mesh-chunks"
@@ -284,26 +277,21 @@ def handle_remesh(table_id):
                 retry = Retry(max=3, interval=[60, 60, 60])
                 queue_name = "mesh-chunks-low-priority"
             q = Queue(queue_name, retry=retry, default_timeout=1200)
-            task = q.enqueue(meshing_tasks.remeshing, table_id, 
-                             new_lvl2_ids)
+            task = q.enqueue(meshing_tasks.remeshing, table_id, new_lvl2_ids)
 
-        response_object = {
-            "status": "success",
-            "data": {
-                "task_id": task.get_id()
-            }
-        }
-        
+        response_object = {"status": "success", "data": {"task_id": task.get_id()}}
+
         return jsonify(response_object), 202
     else:
         new_lvl2_ids = np.array(new_lvl2_ids, dtype=np.uint64)
         cg = app_utils.get_cg(table_id)
-        
+
         if len(new_lvl2_ids) > 0:
-            t = threading.Thread(target=_remeshing, 
-                                 args=(cg.get_serialized_info(), new_lvl2_ids))
+            t = threading.Thread(
+                target=_remeshing, args=(cg.get_serialized_info(), new_lvl2_ids)
+            )
             t.start()
-    
+
         return Response(status=202)
 
 
@@ -326,5 +314,5 @@ def _remeshing(serialized_cg_info, lvl2_nodes):
         cv_sharded_mesh_dir=cv_mesh_dir,
         cv_unsharded_mesh_path=cv_unsharded_mesh_path,
     )
-    
+
     return Response(status=200)
